Fix the last three type errors (dialog conversion, Float/Double peak)
  * VoiceRemote: .result(dialog:) takes an IntentDialog; a bare Text does not
    convert, so the two runtime-computed replies now use
    IntentDialog(content: Text(…)) the way the literal ones use
    IntentDialog(stringLiteral: …).
  * VocalStudio: the gate meter compared Double against Float, which the
    type-checker would not even resolve the literals for. `peak` is a Double
    now (it only ever feeds dB maths and a 0.12 threshold), which removes both
    the comparison error and the Float(1e-6) gymnastics next to it.

+++ TrollMusicApp/TrollMusicApp/VocalStudio.swift
@@ -605,7 +605,7 @@ final class KaraokeRecorder: ObservableObject {
         let sr = format.sampleRate
         // One-pole high-pass at 90 Hz (removes desk rumble and plosive boom).
         let a = 1 - exp(-2 * Double.pi * 90 / sr)
-        var peak: Float = 0
+        var peak: Double = 0
         let ch = Int(format.channelCount)
 
         for c in 0..<min(ch, 2) {
@@ -615,12 +615,12 @@ final class KaraokeRecorder: ObservableObject {
                 hpInput += a * (x - hpInput)
                 let y = x - hpInput                      // high-passed
                 let ay = abs(y)
-                if ay > peak { peak = Float(ay) }
-                let gated = gateEnabled ? (Double(peak) > 0.12 ? y * Double(boost) : 0) : y * Double(boost)
+                if ay > peak { peak = ay }
+                let gated = gateEnabled ? (peak > 0.12 ? y * Double(boost) : 0) : y * Double(boost)
                 p[i] = max(-0.95, min(0.95, Float(gated)))
             }
         }
-        let lvl = peak > 0 ? min(1, Double(20 * log10(max(Float(1e-6), peak)) + 55) / 55) : 0
+        let lvl = peak > 0 ? min(1, (20 * log10(max(1e-6, peak)) + 55) / 55) : 0
         gateHold = max(lvl > 0.10 ? 1.0 : gateHold - Double(n) / sr * 4, 0)
 
         if let f = file {
